+++ src/_internals/tests.js
@@ -1,6 +1,4 @@
-import {
-  extend, generateDate, getCardType, isObject, luhnChk, matchesCardType, normalizeDate, organizeResults, validateType
-} from './index';
+import { extend, getCardType, isObject, organizeResults } from './index';
 import test from 'ava';
 
 test('Testing extend functionality', t => {
@@ -12,13 +10,6 @@ test('Testing extend functionality', t => {
   });
 });
 
-test('Testing generateDate functionality', t => {
-  const results = generateDate();
-  const dateObj = new Date();
-
-  t.deepEqual(results, new Date(`${dateObj.getMonth() + 1}/1/${dateObj.getFullYear()}`));
-});
-
 test('Testing getCardType functionality', t => {
   const results = getCardType('4122027811098688');
 
@@ -28,24 +19,6 @@ test('Testing getCardType functionality', t => {
 test('Testing isObject functionality', t => {
   t.truthy(isObject({}));
   t.falsy(isObject([]));
-});
-
-test('Testing luhnCheck functionality', t => {
-  t.truthy(luhnChk('4122027811098688'));
-  t.falsy(luhnChk('4'));
-});
-
-test('Testing matchesCardType functionality', t => {
-  t.truthy(matchesCardType('333', { info: 'visa' }));
-  t.truthy(matchesCardType('3333', { info: 'amex' }));
-  t.falsy(matchesCardType('333', { info: 'amex' }));
-  t.falsy(matchesCardType('3333', { info: 'visa' }));
-});
-
-test('Testing normalizeDate functionality', t => {
-  t.is(normalizeDate('1/19'), '1/1/2019');
-  t.is(normalizeDate('1 / 19'), '1/1/2019');
-  t.is(normalizeDate('1-19'), '1/1/2019');
 });
 
 test('Testing organizeResults functionality', t => {
@@ -70,10 +43,4 @@ test('Testing organizeResults functionality', t => {
     cvnType: 'norm',
     expired: false
   });
-});
-
-test('Test validateType functionality', t => {
-  t.is(validateType('1/19'), 'validDate');
-  t.is(validateType('4122027811098688'), 'validNumber');
-  t.is(validateType('333'), 'validCVN');
 });